+++ src/gpp_client/coordinator/__init__.py
@@ -1,0 +1,3 @@
+from .base import BaseCoordinator
+
+__all__ = ["BaseCoordinator"]
