+++ src/gpp_client/directors/scheduler/coordinators/__init__.py
@@ -1,0 +1,3 @@
+from .program import ProgramCoordinator
+
+__all__ = ["ProgramCoordinator"]
